+++ src/NzbDrone.Core/Authentication/User.cs
@@ -14,6 +14,9 @@ namespace NzbDrone.Core.Authentication
         public bool IsActive { get; set; }
         public DateTime CreatedAt { get; set; }
         public DateTime? LastLogin { get; set; }
+        public int LoginCount { get; set; }
+        public int WizardAutoShownCount { get; set; }
+        public bool WizardAutoDisabled { get; set; }
         public string PreferredQualityMedia { get; set; }
         public string ResetToken { get; set; }
         public DateTime? ResetTokenExpiration { get; set; }